pybind11: Commented out some code in pybind11
There is a problem when virtual method that is defined in Python calls
back to C++ and the code in C++ calls back to this overridden method in
Python. There is explicit code that prohibits that and I don't really
know why. After commenting the code out, everything seems to work now
and as we would expect.