utils: Visitee now uses variant instead of optional
In C++ we can distinguish between nullopt and nullptr in
optional<shared_ptr<T>> easily. The problem is that pybind11 translates
both of those into None in Python so we have no longer way to recognize
whether user wanted to keep (nullopt) or delete node (nullptr). With
variant, which had shared_ptr<T> as first, we can easily keep the same
behavior but instead of passing nullptr when delete is required, we just
pass another enum (Visitee::Action::Delete).