types: Removed option for single element sets
The only single element set which is allowed is 'them'. Other sets need
parentheses around it but the problem is what if we have one element
set? Rather than distinguishing it with some wild conditions, it is much
more better to just prohibit them and let developer pass ThemExpression
directly in place of a set.