Simplify usage messages and quitting in Python examples.
sys.argv[0] is the name of the script, so we can use it instead of the
hard-coded name that is prone to become outdated when the script is renamed.

Instead of a print(msg) followed by sys.exit(1), we can use just sys.exit(msg),
which does the same thing but is shorter.