ci: Fix the GithubActions tests on macos-latest
For some reason starting from python3.10.9 the build of yaramod on macos
was not successful. This also affected current master tests. Changing
the build command to use the setuptools expected method of installation
helped.

I had to add a setup.py option that checks the ENV variable to enable
yaramod tests. This is because passing the parameters to `setup.py
build` is not supported anymore and we cannot easily pass
`--with-unit-tests`.