fix: Open json modules in read only mode
We had some cases when the modules were installed with read only
permissions and yaramod would fail to open them.

Also we expect no writes to the modules json.

I removed the `std::ios::out` which opens the file in write mode. The
default configuration is `std::ios_base::in`, which opens the file in
read mode.