docs: Improve release docs
I often forget to bump the version in the docs. So I added this as a
step.

Additionally, after the current steps we don't have the release commit
on the master branch. I am not sure if this was on purpose, but some
older commits end up on master.

It can be because, if you push the commit without first releasing,
GitHub will reject the push, because we have it set up to accept only
tested commits. But after the release is done, we can push the commit to
master since it is considered tested and passing from the release.

If we don't have the commits in master, it can be easy to accidentally
bump the current version and release it without knowing there is already
newer version that what is specified in master.