fix(Rule): Delete tags from the rule
Currently, after we delete tags from the rule, we would still have a `:`
with empty tag list. Which is a incorrect syntax. For example:
```yara
rule foo : bar {
    condition:
        true
}
```
will become (which should not happen)
```yara
rule foo : {
    condition:
        true
}
```

This patch refactors the `Rule::setTags` method to always delete the old
tags if present. And then add new tags only if necessary.