Added deferred include parsing mode + fixed incomplete mode when import of a module is missing
This change adds new deferred include parsing mode which is an extension
of incomplete parsing mode. On top of allowing rules to use undefined
symbols, it additionally does not resolve include directives and let's
user to handle them.

Incomplete mode was also broken for a long time by not allowing to parse
incomplete rules when they used a module for which the definition was
missing. This change forces the import of said module.